TradCreadesFaltaAcabar-les

+++ routes/web.php
@@ -11,27 +11,25 @@ Route::get('/comandes', [ComandaController::class, 'index'])
     ->middleware('auth')
     ->name('comandes');
 
-Route::get('/administrar/productes', [ProducteController::class, 'llistar'])->name('productes');
-Route::get('/producte/{producte}/edit', [ProducteController::class, 'edit'])->name('producte.edit');
-Route::delete('/producte/{producte}', [ProducteController::class, 'destroy'])->name('producte.destroy');
-Route::put('/producte/{producte}/update', [ProducteController::class, 'update'])->name('producte.update');
-Route::get('/producte/create', [ProducteController::class, 'create'])->name('producte.create');
-Route::post('/producte/store', [ProducteController::class, 'store'])->name('producte.store');
+Route::get('/administrar/productes', [ProducteController::class, 'llistar'])->middleware(['auth'])->name('productes');
+Route::get('/producte/{producte}/edit', [ProducteController::class, 'edit'])->middleware(['auth'])->name('producte.edit');
+Route::delete('/producte/{producte}', [ProducteController::class, 'destroy'])->middleware(['auth'])->name('producte.destroy');
+Route::put('/producte/{producte}/update', [ProducteController::class, 'update'])->middleware(['auth'])->name('producte.update');
+Route::get('/producte/create', [ProducteController::class, 'create'])->middleware(['auth'])->name('producte.create');
+Route::post('/producte/store', [ProducteController::class, 'store'])->middleware(['auth'])->name('producte.store');
 
-Route::get('/administrar/categories', [CategoriaController::class, 'llistar'])->name('categories');
-Route::get('/categories/create', [CategoriaController::class, 'create'])->name('categoria.create');
-Route::post('/categories/store', [CategoriaController::class, 'store'])->name('categoria.store');
-Route::get('/categories/{categoria}/edit', [CategoriaController::class, 'edit'])->name('categoria.edit');
-Route::put('/categories/{categoria}/update', [CategoriaController::class, 'update'])->name('categoria.update');
-Route::delete('/categories/{categoria}', [CategoriaController::class, 'destroy'])->name('categoria.destroy');
+Route::get('/administrar/categories', [CategoriaController::class, 'llistar'])->middleware(['auth'])->name('categories');
+Route::get('/categories/create', [CategoriaController::class, 'create'])->middleware(['auth'])->name('categoria.create');
+Route::post('/categories/store', [CategoriaController::class, 'store'])->middleware(['auth'])->name('categoria.store');
+Route::get('/categories/{categoria}/edit', [CategoriaController::class, 'edit'])->middleware(['auth'])->name('categoria.edit');
+Route::put('/categories/{categoria}/update', [CategoriaController::class, 'update'])->middleware(['auth'])->name('categoria.update');
+Route::delete('/categories/{categoria}', [CategoriaController::class, 'destroy'])->middleware(['auth'])->name('categoria.destroy');
 
-Route::post('/afegir-carro/{producte}', [CarritoController::class, 'afegir'])->name('afegir.carro');
-Route::post('/buidar-carro', [CarritoController::class, 'buidar'])->name('buidar.carro');
-Route::post('/comprar-carro/{metode}', [CarritoController::class, 'comprar'])->name('comprar.carro');
+Route::post('/afegir-carro/{producte}', [CarritoController::class, 'afegir'])->middleware(['auth'])->name('afegir.carro');
+Route::post('/buidar-carro', [CarritoController::class, 'buidar'])->middleware(['auth'])->name('buidar.carro');
+Route::post('/comprar-carro/{metode}', [CarritoController::class, 'comprar'])->middleware(['auth'])->name('comprar.carro');
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::get('/dashboard', function () {return view('dashboard');})->middleware(['auth', 'verified'])->name('dashboard');
 
 
 Route::get('/', [CategoriaController::class, 'index'])->name('home');
